changed: avoid some warnings in new caching system
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk/XBMC@10651 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/FileSystem/CacheMemBuffer.h
@@ -32,7 +32,7 @@ public:
 
     virtual int WriteToCache(const char *pBuffer, size_t iSize) ;
     virtual int ReadFromCache(char *pBuffer, size_t iMaxSize) ;
-    virtual __int64 WaitForData(unsigned int iMinAvail, unsigned int iMillis) ;
+    virtual __int64 WaitForData(__int64 iMinAvail, unsigned int iMillis) ;
 
     virtual __int64 Seek(__int64 iFilePosition, int iWhence) ;
 	virtual void Reset(__int64 iSourcePosition) ;
